[ExternalLink] Move under components/material
This component is used by two components now so it should be moved up
